requirements.txt
mysql-connector-python>=9.2.0
python-dotenv>=1.0.1
openai>=1.69.0
nonebot-plugin-alconna>=0.56.2